is_verified_user for check
Added check if ticket holder is current user
