+++ app/templates/pdf/ticket_attendee.html
@@ -7,10 +7,12 @@
             size: a4 portrait;
             background-image: url('{{ base_dir }}/static/data/ticket-trans-notext.png');
             margin: 0;
+
             @frame col1_frame {             /* Content frame 1 */
                 left: 80pt;  top: 30pt;
                 height: 250pt; width: 300pt;
             }
+
             @frame qrcode {
                 left: 455pt;  top: 50pt;
                 width: 100pt; height: 120pt;
@@ -20,16 +22,21 @@
                 width: 100pt; height: 30pt;
                 -pdf-frame-content: number_content;
             }
+
+
              @frame col2_frame {
                 left: 440pt; top: 170pt;
             }
         }
+
         #qrcode {
             -pdf-frame-break: true;
         }
+
         .label {
             color: #858585;
         }
+
     </style>
 </head>
 <body>
